% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.r
\name{estimate_beta}
\alias{estimate_beta}
\title{Estimate beta function}
\usage{
estimate_beta(y, X)
}
\arguments{
\item{y}{A vector of n observations.}

\item{X}{A n x k matrix.}
}
\value{
The estimates of the linear regression coefficients.
}
\description{
Estimate beta function
}
\examples{
vec <- c(1, 5, 6)
mat <- matrix( c(5, 1, 0, 3,-1, 2, 4, 0,-1), nrow=3, byrow=TRUE)
estimate_beta(vec, mat)
}
