% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_theme.r
\name{my_theme}
\alias{my_theme}
\title{My theme function}
\usage{
my_theme(background = "darkblue", line = "dashed", panel = "lightskyblue")
}
\arguments{
\item{background}{A color to set the theme's background.}

\item{line}{A style for the lines around the border.}

\item{panel}{A color to set the theme's panel.}
}
\value{
A ggplot theme.
}
\description{
My theme function
}
\examples{
my_theme("red", "longdash", "yellow")
}
